+++ src/roles/role.upgrader.js
@@ -2,7 +2,7 @@ module.exports = {
 
   run: creep => {
 
-    if (creep.recycleAt(20)) return;
+    if (creep.recycleAt()) return;
 
     if (creep.isCharged()) {